Started adding new functionality to ground truth interface
Now started adding new functionality to the ground truth
preparing interface.
The idea is that it shall permit us to mark text regions.
And that each point should also belong to a region.
By default all points belong to root region.
If non other regions are specified, then the image can be
treated as an example of 1, that is directly usable for line
segmentation. Else it needs of text region detection.