mark columns v2 started
Mark columns v2 function is written.
The idea now is to find the lowest some by using windows. We
save the indices of the lowest sum to backtrack as well.
Backtrack matrix have 6 elements in a row:
MinRow, minCol, localMinEnergy, currentRow, currentCol, currentPathEnergy

MinRow, and minCol means the indices of the pixel that is around the
currentRow and currentCol with lowest energy.

The algorithm for backtracking is simple.
At the beginning we obtain the lowest energy pixel at the bottom of the
energy table matrix.
Then we find the rows in backtrack matrix that keeps the indices coming
from the energy table as its MinRow and minCol.
From those rows we select the one with the lowest currentPathEnergy.
We use the currentRow and currentCol values of the obtained row to repeat the
process.
We terminate once we arrive at the top row of the energy table.